feat: Fix IsPinch calculation in CharacterInstanceStatus
The calculation for the `IsPinch` property in the `CharacterInstanceStatus` class has been fixed to correctly calculate the pinch threshold. Previously, the `HitPoint` and `HitPointMax` values were divided as integers, resulting in incorrect comparisons. This change ensures that the pinch threshold is accurately determined based on the ratio of the current hit points to the maximum hit points.

+++ Assets/SoulRPG/Scripts/CharacterControllers/CharacterInstanceStatus.cs
@@ -34,7 +34,7 @@ namespace SoulRPG
         public ReadOnlyReactiveProperty<int> ExperienceAsObservable() => experienceReactiveProperty;
         public int Experience => experienceReactiveProperty.Value;
 
-        public bool IsPinch => (HitPoint / HitPointMax) < TinyServiceLocator.Resolve<GameRule>().PinchThreshold;
+        public bool IsPinch => ((float)HitPoint / HitPointMax) < TinyServiceLocator.Resolve<GameRule>().PinchThreshold;
 
         public CharacterInstanceStatus(Character character, CharacterGrowthParameter growthParameter)
         {
